remove lots of commented out code from GameWindow file as it was all out of date, 	handle embed segue for result navigation
remove lots of nslog statements which are now not useful

in fbdatacontroller.h add sessionResults object to eventually support storing session results separately.
	add a mutable array property to store all the different results objects that have been encountered.
	add a results index property to keep track of the current user's results.  (These are the only ones that can currently be displayed).
	make the addResultsForCurrentUser publicly accessible so it can be called in the correct place.

In fbdatacontroller.m initialize the allResults object with a blank Mutable Array
	create addResultsForCurrentUser function to handle finding the current user's previous results if they exist and if not adding a blank results object with their name.
	add necessary checks to addResultsObj to avoid nil exception error.  Problem occurred where objects weren't correctly initialized.
	update incrementSuccesses, incrementAttempts, getSessionAttempts, and getSessionSuccesses to reflect new storage style of results objects.
	change the writing of a file to correctly store all the results objects that exist in allResults in a sane way so as to retrieve them later.

In frgLoggedInBaseViewController.m, move improper initialization of data controller object to the viewDidLoad section.
	make request in viewDidAppear for the current user's name and then update this in the data controller
	remove some now extraneous debugging code. Add some other now necessary debugging code.

in resultsObj.h file, add property to store the username associated with the object.
	add functions to return the attempts made and successes achieved.

in resultsObj.m, add initialization of userName to the initWithValue function.
	update the initWithCoder to support proper decoding functionality.
	update the encodeWithCoder to support proper encoding functionality.

In storyboard, add the container view and embed segue to support the 'potentially' new results navigation table view.
add two new eatableview files to support result navigation.  The idea for result navigation is to have the option of viewing current session data or all previous data.